+++ c/py_types.h
@@ -58,6 +58,7 @@ PyObject* int_to_py(int32_t x);
 PyObject* int_to_py(int64_t x);
 PyObject* float_to_py(float x);
 PyObject* float_to_py(double x);
+PyObject* string_to_py(const CString& x);
 
 
 #endif